Safari-611.1.10/LayoutTests/fast/selectors/nth-last-child-chained-expected.txt

Test chaining many :nth-last-child() selectors. This verifies register allocation is correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-last-child(-n+21):nth-last-child(odd):nth-last-child(-2n+21):nth-last-child(3n+1):nth-last-child(-3n+22):nth-last-child(4n+1):nth-last-child(-5n+51)"
PASS document.querySelectorAll('testcase:nth-last-child(-n+21):nth-last-child(odd):nth-last-child(-2n+21):nth-last-child(3n+1):nth-last-child(-3n+22):nth-last-child(4n+1):nth-last-child(-5n+51)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(-n+21):nth-last-child(odd):nth-last-child(-2n+21):nth-last-child(3n+1):nth-last-child(-3n+22):nth-last-child(4n+1):nth-last-child(-5n+51)')[0].id is "testcase20"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(10, 100, 200)"
PASS document.querySelectorAll('div>testcase:nth-last-child(-n+21):nth-last-child(odd):nth-last-child(-2n+21):nth-last-child(3n+1):nth-last-child(-3n+22):nth-last-child(4n+1):nth-last-child(-5n+51)>div>div>div span').length is 1

Testing ":nth-last-child(2n-200):nth-last-child(3n-1):nth-last-child(4n-38):nth-last-child(5n-98)"
PASS document.querySelectorAll('testcase:nth-last-child(2n-200):nth-last-child(3n-1):nth-last-child(4n-38):nth-last-child(5n-98)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(2n-200):nth-last-child(3n-1):nth-last-child(4n-38):nth-last-child(5n-98)')[0].id is "testcase19"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(2n-200):nth-last-child(3n-1):nth-last-child(4n-38):nth-last-child(5n-98)>div>div>div span').length is 1

Testing ":nth-last-child(even):nth-last-child(3n):nth-last-child(4n-2):nth-last-child(5n+1)"
PASS document.querySelectorAll('testcase:nth-last-child(even):nth-last-child(3n):nth-last-child(4n-2):nth-last-child(5n+1)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(even):nth-last-child(3n):nth-last-child(4n-2):nth-last-child(5n+1)')[0].id is "testcase15"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(even):nth-last-child(3n):nth-last-child(4n-2):nth-last-child(5n+1)>div>div>div span').length is 1

Testing ":nth-last-child(-n+9000):nth-last-child(-2n+683):nth-last-child(-3n+31):nth-last-child(-4n+47):nth-last-child(-5n+107):nth-last-child(-6n+73):nth-last-child(-7n+70)"
PASS document.querySelectorAll('testcase:nth-last-child(-n+9000):nth-last-child(-2n+683):nth-last-child(-3n+31):nth-last-child(-4n+47):nth-last-child(-5n+107):nth-last-child(-6n+73):nth-last-child(-7n+70)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(-n+9000):nth-last-child(-2n+683):nth-last-child(-3n+31):nth-last-child(-4n+47):nth-last-child(-5n+107):nth-last-child(-6n+73):nth-last-child(-7n+70)')[0].id is "testcase14"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(-n+9000):nth-last-child(-2n+683):nth-last-child(-3n+31):nth-last-child(-4n+47):nth-last-child(-5n+107):nth-last-child(-6n+73):nth-last-child(-7n+70)>div>div>div span').length is 1

Testing ":nth-last-child(odd):nth-last-child(3n+1):nth-last-child(4n-1):nth-last-child(5n+4)"
PASS document.querySelectorAll('testcase:nth-last-child(odd):nth-last-child(3n+1):nth-last-child(4n-1):nth-last-child(5n+4)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(odd):nth-last-child(3n+1):nth-last-child(4n-1):nth-last-child(5n+4)')[0].id is "testcase2"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(odd):nth-last-child(3n+1):nth-last-child(4n-1):nth-last-child(5n+4)>div>div>div span').length is 1

Testing ":nth-last-child(-n+1500):nth-last-child(2n+1):nth-last-child(-3n+45):nth-last-child(4n+3):nth-last-child(-5n+1545)"
PASS document.querySelectorAll('testcase:nth-last-child(-n+1500):nth-last-child(2n+1):nth-last-child(-3n+45):nth-last-child(4n+3):nth-last-child(-5n+1545)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(-n+1500):nth-last-child(2n+1):nth-last-child(-3n+45):nth-last-child(4n+3):nth-last-child(-5n+1545)')[0].id is "testcase6"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(-n+1500):nth-last-child(2n+1):nth-last-child(-3n+45):nth-last-child(4n+3):nth-last-child(-5n+1545)>div>div>div span').length is 1

Testing ":nth-last-child(n+2):nth-last-child(-n+6000):nth-last-child(n-2):nth-last-child(2n):nth-last-child(-2n+20):nth-last-child(2n+2):nth-last-child(3n+1):nth-last-child(-3n+49):nth-last-child(3n-92):nth-last-child(4n+2):nth-last-child(-4n+50)"
PASS document.querySelectorAll('testcase:nth-last-child(n+2):nth-last-child(-n+6000):nth-last-child(n-2):nth-last-child(2n):nth-last-child(-2n+20):nth-last-child(2n+2):nth-last-child(3n+1):nth-last-child(-3n+49):nth-last-child(3n-92):nth-last-child(4n+2):nth-last-child(-4n+50)').length is 1
PASS document.querySelectorAll('testcase:nth-last-child(n+2):nth-last-child(-n+6000):nth-last-child(n-2):nth-last-child(2n):nth-last-child(-2n+20):nth-last-child(2n+2):nth-last-child(3n+1):nth-last-child(-3n+49):nth-last-child(3n-92):nth-last-child(4n+2):nth-last-child(-4n+50)')[0].id is "testcase11"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-last-child(n+2):nth-last-child(-n+6000):nth-last-child(n-2):nth-last-child(2n):nth-last-child(-2n+20):nth-last-child(2n+2):nth-last-child(3n+1):nth-last-child(-3n+49):nth-last-child(3n-92):nth-last-child(4n+2):nth-last-child(-4n+50)>div>div>div span').length is 1

PASS successfullyParsed is true

TEST COMPLETE
